refactor: clean and optimize code
