+++ src/components/portal/auth/styled.js
@@ -73,14 +73,13 @@ export const SButton = styled.button`
     background-color: white;
     transition: 0.5s ease-out;
   }
-`
-
-export const SDesabledButton = styled.button`
-  text-align: center;
-  font-size: 16px;
-  border: 1px solid rgb(228, 224, 224);
-  background-color: white;
-  padding: 5px 25px;
-  color: #771700;
-  margin: 30px 0px;
+  :disabled {
+    text-align: center;
+    font-size: 16px;
+    border: 1px solid rgb(228, 224, 224);
+    background-color: white;
+    padding: 5px 25px;
+    color: #771700;
+    margin: 30px 0px;
+  }
 `
