+++ src/components/portal/cards/list/styled.js
@@ -9,62 +9,11 @@ export const SCard = styled.div`
   background-color: #fff;
   box-shadow: -1px 1px 0 #dcdcdc;
   text-align: justify;
-  margin: 50px 160px;
+  margin: 50px auto;
   :hover {
     box-shadow: 5px 10px 20px 1px rgba(0, 0, 0, 0.253) !important;
     transition: 1s;
     opacity: 0.5;
-  }
-  @media screen and (max-width: 1700px) {
-    margin: 50px 120px;
-  }
-
-  @media screen and (max-width: 1500px) {
-    margin: 50px 70px;
-  }
-
-  @media screen and (max-width: 1240px) {
-    margin: 50px 60px;
-  }
-
-  @media screen and (max-width: 1200px) {
-    margin: 50px 140px;
-  }
-
-  @media screen and (max-width: 1100px) {
-    margin: 50px 115px;
-  }
-
-  @media screen and (max-width: 1000px) {
-    margin: 50px 85px;
-  }
-
-  @media screen and (max-width: 930px) {
-    margin: 50px 70px;
-  }
-
-  @media screen and (max-width: 860px) {
-    margin: 50px 50px;
-  }
-
-  @media screen and (max-width: 770px) {
-    margin: 50px 230px;
-  }
-
-  @media screen and (max-width: 625px) {
-    margin: 50px 150px;
-  }
-
-  @media screen and (max-width: 525px) {
-    margin: 50px 80px;
-  }
-
-  @media screen and (max-width: 487px) {
-    margin: 50px 45px;
-  }
-
-  @media screen and (max-width: 380px) {
-    margin: 50px 30px;
   }
 `
